Implemented edit and reply.

+++ migrations/2018_06_30_113500_create_comments_table.php
@@ -18,7 +18,7 @@ class CreateCommentsTable extends Migration
             $table->unsignedBigInteger('commenter_id');
             $table->morphs('commentable');
             $table->text('comment');
-            $table->unsignedInteger('child_id')->nullable();
+            $table->unsignedBigInteger('child_id')->nullable();
             $table->foreign('child_id')->references('id')->on('comments')->onDelete('cascade');
             $table->timestamps();
         });
